Rename `BloodSugarValidation#Result` to `ValidationResult`
This is in preparation for a future commit where we will move the
class to its own file.
